requirements.txt
wheel
asgiref             3.4.1  
Django              3.2.9  
django-crispy-forms 1.13.0 
django-reversion    4.0.1  
pkg-resources       0.0.0  
pytz                2021.3 
setuptools          44.0.0 
sqlparse            0.4.2  
